+++ scripts/core/helpers/network.ts
